fix: Correct primary screen detection
In KDE, the index of the primary screen always appears to be 0. Based on
actual tests, it functions correctly regardless of whether the primary
monitor is positioned on the left or right.

https://github.com/nclarius/KWin-window-positioning-scripts/blob/fe9d1641982912075f9debe68cc6b9afde7292d5/always-open-on-primary-screen/contents/code/main.js

The previous approach expected the activeScreen to be the primary at the
time the script was activated. This method could lead to misdetections,
which I have experienced.